feat(wasm): split crate to allow compiling as a library
