+++ src/lib.rs
@@ -1,0 +1,3 @@
+mod instructions;
+pub mod server;
+
